+++ src/plugins/piral-breadcrumbs/src/Breadcrumbs.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react';
+import { useLocation, useParams } from 'react-router';
+import { PiralBreadcrumbsContainer, PiralBreadcrumbItem } from './components';
+import { useBreadcrumbs } from './useBreadcrumbs';
+
+export const Breadcrumbs: React.FC = () => {
+  const location = useLocation();
+  const params = useParams();
+  const breadcrumbs = useBreadcrumbs(location.pathname);
+  const currentIndex = breadcrumbs.length - 1;
+
+  const children = breadcrumbs.map(({ settings }, i) => {
+    const { title, path, ...props } = settings;
+    const key = `bc_${i}_${settings.path}`;
+    const current = i === currentIndex;
+    const computedPath = path.replace(/:([A-Za-z0-9_]+)/g, (s, id) => params[id] ?? s);
+
+    return (
+      <PiralBreadcrumbItem key={key} current={current} path={computedPath} {...props}>
+        {title}
+      </PiralBreadcrumbItem>
+    );
+  });
+
+  return <PiralBreadcrumbsContainer children={children} />;
+};
+Breadcrumbs.displayName = 'Breadcrumbs';