+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added `piral-cli-webpack5` to support bundling with Webpack 5 (#313)
 * Added the `packageOverrides` field to override *package.json* values of pilets (#330)
 * Added special tagged comments for additional modifications of the `PiletApi` (#332)
+* Added `piral-breadcrumbs` for bringing managing breadcrumbs
 
 ## 0.12.3 (October 10, 2020)